Add some new properties to the CoapClient class
Two new properties added tothe CoapClient class to make life easier when
programming with these classes.

UriPath - create a property that holds a relative path to the main URI
in Uri.  This means that one can create the class with just the server
name and then use this to access different resources on the sever

UriQuery - create a property that holds a query string to be added to
the Uri when the message is sent.