Multicast testing and fixes
The multicast code did not quite work as it was supposed to.  Do rewrites so that sockets are no going to be re-used for two different servers.  Disabled the dual stack mode on all of the sockets because it has the problem that the multicast address returned is wrong for an v4 address.